requirements.txt
django==3.2.7
djangorestframework
django-cors-headers
pillow
django-environ
drf_yasg
djangorestframework-simplejwt
gunicorn==20.1.0
psycopg2-binary
# redis==4.3.1
# celery==5.2.6
